Moved skyblockdata package to data package
