+++ src/main/java/me/partlysanestudios/partlysaneskies/data/skyblockdata/SkyblockPlayer.java
@@ -3,7 +3,7 @@
 // See LICENSE for copyright and license notices.
 //
 
-package me.partlysanestudios.partlysaneskies.skyblockdata;
+package me.partlysanestudios.partlysaneskies.data.skyblockdata;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
